// src/survey/form/model/QuestionItemType.ts
export enum QuestionItemType {
  Criterion = 'Criterion', // 척도
  Choice = 'Choice', // 객관식
  Essay = 'Essay', // 주관식
  Date = 'Date', // 날짜형식
  Boolean = 'Boolean', // YesNo
  Matrix = 'Matrix', // 행렬
  Review = 'Review', // 리뷰형
  ChoiceFixed = 'ChoiceFixed', // 아이콘형
}